requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.7.0
    # via pydantic
backoff==2.2.1
    # via rudder-sdk-python
cachetools==5.5.0
    # via google-auth
certifi==2024.8.30
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via typer
colorama==0.4.6
    # via typer
dbtc==0.11.4
    # via -r requirements.in
google-api-core==2.20.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-core
google-auth==2.35.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-core
google-cloud-bigquery==3.26.0
    # via -r requirements.in
google-cloud-core==2.4.1
    # via google-cloud-bigquery
google-crc32c==1.6.0
    # via google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-bigquery
googleapis-common-protos==1.65.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.66.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.66.1
    # via google-api-core
idna==3.10
    # via requests
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
monotonic==1.6
    # via rudder-sdk-python
numpy==1.26.4
    # via
    #   pandas
    #   pyarrow
packaging==24.1
    # via google-cloud-bigquery
pandas==2.1.4
    # via dbtc
proto-plus==1.24.0
    # via google-api-core
protobuf==5.28.2
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyarrow==15.0.2
    # via dbtc
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.9.2
    # via dbtc
pydantic-core==2.23.4
    # via pydantic
pygments==2.18.0
    # via rich
python-dateutil==2.9.0.post0
    # via
    #   google-cloud-bigquery
    #   pandas
    #   rudder-sdk-python
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.2
    # via pandas
requests==2.32.3
    # via
    #   dbtc
    #   google-api-core
    #   google-cloud-bigquery
    #   rudder-sdk-python
rich==13.8.1
    # via typer
rsa==4.9
    # via google-auth
rudder-sdk-python==1.0.6
    # via dbtc
shellingham==1.5.4
    # via typer
six==1.16.0
    # via
    #   python-dateutil
    #   rudder-sdk-python
typer==0.9.4
    # via dbtc
typing-extensions==4.12.2
    # via
    #   pydantic
    #   pydantic-core
    #   typer
tzdata==2024.2
    # via pandas
urllib3==1.26.20
    # via requests
